requirements.txt
asgiref==3.11.1
attrs==26.1.0
dj-rest-auth==7.1.1
Django==4.2
django-cors-headers==3.10.0
djangorestframework==3.16.1
drf-spectacular==0.21.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.1.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
